project/requirements/base.txt
# base.txt is a pip requirements file which describes the necessary
# distributions required to run this project. If you need something that is
# project specific not listed here use project.txt. You can, of course, update
# versions of distributions here if needed.

--extra-index-url=http://dist.pinaxproject.com/dev/

#Django==1.2.1
django-debug-toolbar==0.8.3
django-staticfiles==0.2.0
